Rename 'evaluationengine' package to 'engine'
