+++ rulette-core/src/main/java/com/github/kislayverma/rulette/core/RuleSystem.java
@@ -3,8 +3,8 @@ package com.github.kislayverma.rulette.core;
 import com.github.kislayverma.rulette.core.dao.DataSource;
 import com.github.kislayverma.rulette.core.dao.RuleSystemDao;
 import com.github.kislayverma.rulette.core.dao.impl.mysql.RuleSystemDaoImpl;
-import com.github.kislayverma.rulette.core.evaluationengine.IEvaluationEngine;
-import com.github.kislayverma.rulette.core.evaluationengine.impl.trie.TrieBasedEvaluationEngine;
+import com.github.kislayverma.rulette.core.engine.IEvaluationEngine;
+import com.github.kislayverma.rulette.core.engine.impl.trie.TrieBasedEvaluationEngine;
 import com.github.kislayverma.rulette.core.metadata.RuleSystemMetaData;
 import com.github.kislayverma.rulette.core.metadata.RuleSystemMetaDataFactory;
 import com.github.kislayverma.rulette.core.rule.Rule;
